Temporarily patched resizing issues
Recent updates to iPlug2 allow non-integer window scaling to adapt the window size to high-DPI displays, whereas previously the scaling factor of 1.25 for my screen was truncated to 1. The automatic DPI scaling, however, causes display issues for Tablitsa because the UI was not designed with it in mind, so I've commented out the problematic lines for now.